% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams-AS7343.r
\docType{data}
\name{ams_AS7331.spct}
\alias{ams_AS7331.spct}
\title{ams-AS7331 digital ultraviolet sensor}
\format{
A \code{response_spct} object containing 3 spectra in long form.
}
\usage{
ams_AS7331.spct
}
\description{
AS7331 digital ultraviolet sensor combining an array of photodiodes with
interference filters deposited directly on the sensor chip. The sensor
digitises the readings and communicates through I2C interface. In addition
to retrieving readings through I2C the analogue gain and integration time
used an be programmed.
}
\note{
Digitized with 'DigitizeIt' from manufacturers' electronic component data
sheet. This are approximate data, both because of the digitizing process,
and because they are typical values. Individual sensor units are expected
to differ to some degree in spectral response.

Manufacturer: ams-OSRAM AG, Austria.
}
\examples{

levels(ams_AS7331.spct[[getIdFactor(ams_AS7331.spct)]])
summary(ams_AS7331.spct, expand = "collection")

}
\references{
Data sheet for AS7331, ams-OSRAM.
}
\keyword{datasets}
