% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams-AS7343.r
\docType{data}
\name{ams_TSL2591.spct}
\alias{ams_TSL2591.spct}
\title{ams-TSL2591 ambient light sensor}
\format{
A \code{response_spct} object containing 2 spectra in long form.
}
\usage{
ams_TSL2591.spct
}
\description{
TSL2591 digital ambient light sensor combining two photodiodes, one bare
and one filtered. The sensor digitises the readings and communicates through
I2C interface.
}
\note{
Digitized with 'DigitizeIt' from manufacturers' electronic component data
sheet. This are approximate data, both because of the digitizing process,
and because they are typical values. Individual sensor units are expected
to differ to some degree in spectral response.

Manufacturer: ams-OSRAM AG, Austria.
}
\examples{

levels(ams_TSL2591.spct[[getIdFactor(ams_TSL2591.spct)]])
summary(ams_TSL2591.spct, expand = "collection")

}
\references{
Data sheet for TSL2591, ams-OSRAM.
}
\keyword{datasets}
