% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytik-sensors.r
\docType{data}
\name{analytik_sensors}
\alias{analytik_sensors}
\title{Analytik-Jena sensors.}
\format{
An object of class \code{character} of length 0.
}
\usage{
analytik_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
UVX-25, UVX-31 and UVX-36 sensor for UV-C, UV-B and UV-A radiation,
respectively. Detection system includes optical filters and a Si photodiode.

Manufacturer: Analytik Jena US LLC, Upland, CA, USA (former UVP)
\url{https://www.analytik-jena.com/} or \url{https://www.uvp.com/}.
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{

analytik_sensors

}
\references{
Product manual: 'UVP UVX Radiometer: Instruction Guide' (figures 6, 7 and 8).
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
