% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular-response..R
\name{angular_response}
\alias{angular_response}
\title{Angular response}
\usage{
angular_response(
  elevation.angle = 90,
  geometry = "cosine",
  zenith.angle = 90 - elevation.angle,
  diameter = NULL
)
}
\arguments{
\item{elevation.angle, zenith.angle}{numeric The elevation angle of
a point light source such as the sun or its zenith angle [degrees].}

\item{geometry}{character The type of entrance optics, one of
\code{"flat disk"}, \code{"cosine"}, \code{"dome"}, \code{"hemisphere"},
\code{"ball"}, or \code{"sphere"}.}

\item{diameter}{numeric The diameter of the entrance optics. If \code{NULL},
the default, a relative value is returned.}
}
\description{
Angular response of idealized entrance optics used in light measurements.
}
\details{
The maximum projected area (\eqn{A_\mathrm{max}}) is always computed
  for a circle of diameter \eqn{d} as \eqn{A_\mathrm{max} = \pi \times d^2 / 4}
  when computing actual projected areas, or set to \eqn{A_\mathrm{max} = 1}
  for computation of relative values.

The \emph{cosine response} for a \emph{flat disk} is computed as
\deqn{A_\mathrm{p} = A_\mathrm{max} \times \cos(z)}

The \emph{hemispherical response} for a \emph{dome} is computed as
\deqn{A_\mathrm{p} = A_\mathrm{max} \times 0.5 \times (1 + \cos(z))}

The \emph{spherical response} for a \emph{"ball"} is computed as
\deqn{A_\mathrm{p} = A_\mathrm{max} \times 1}
}
\examples{
angular_response(45)
angular_response(45, "cosine")
angular_response(45, "dome")
angular_response(45, "sphere")
angular_response(c(0, 30, 60, 90))
angular_response(-c(0, 30, 60, 90))
angular_response(c(0, 30, 60, 90), "dome")
angular_response(c(0, 30, 60, 90), "sphere")

}
\seealso{
\code{\link{all_diffusers}} and \code{\link{diffusers.lst}} for
  data for real sensors and entrance optics.
}
