% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee-sensors.r
\docType{data}
\name{apogee_sensors}
\alias{apogee_sensors}
\title{apogee broadband sensors}
\format{
An object of class \code{character} of length 6.
}
\usage{
apogee_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
apogee SQ-100X-SS: Original Quantum Sensor.

apogee SQ-500-SS: Full-Spectrum Quantum Sensor

apogee SQ-610-SS 400-750nm ePAR Sensor (extended PAR quantum)

apogee SU-200-SS: UV-A Sensor (UV-A, "energy")

apogee S2-131-SS Red and Far-Red Sensor, Red channel, quantum

apogee S2-131-SS Red and Far-Red Sensor, Far-Red channel, quantum

Manufacturer: Apogee Instruments, Inc., Logan, UT, USA.
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{

apogee_sensors

}
\references{
\url{https://www.apogeeinstruments.com/quantum/}
visited on 2023-10-10. (manufacturers product on-line specifications).
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
