% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaT-sensors.r
\docType{data}
\name{deltat_sensors}
\alias{deltat_sensors}
\title{Delta-T sensors}
\format{
An object of class \code{character} of length 1.
}
\usage{
deltat_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
From figure in manufacturer's manual. The Delta-T BF5 in a broadband
diffuse and total irradiance sensor calibrated for photosynthetically
active radiation (PAR). It is based on seven Si photodiodes as a shading
dome with a custom pattern of openings.

Manufacturer: Delta-T Devices Ltd, CAMBRIDGE CB25 0EJ, UK
\url{https://delta-t.co.uk/}
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{
deltat_sensors

}
\references{
Manufacturer's User Manual Version: 1.0 dated Nov 2010
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
