% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusers-lst.r
\docType{data}
\name{diffusers.lst}
\alias{diffusers.lst}
\title{Angular response of sensors}
\format{
A \code{response_mspct} object containing a
\code{response_spct} objects as \emph{named} members.

Each member spectrum contains three variables:
\itemize{
   \item angle.deg (degrees)
   \item response (/1)
   \item response.over.cosine (/1)
}
}
\usage{
diffusers.lst
}
\description{
A collection of angular response data for selected broadband
sensors used for measuring ultraviolet and visible radiation and of cosine
diffusers used with spectrometers. Each data frame in the
collection contains three variables, angle.deg (degrees) at either regular or
irregular intervals, response relative to the maximum (as a fraction of one)
and response relative to a perfect cosine response(as a fraction of one).
Data are either from manufacturer specifications or independent measurements
reported in the scientific literature.
}
\note{
Values are only good as reference, as individual sensors and diffuser
  deviate to a smaller or lager extent from typical or mean responses.
}
\examples{
names(diffusers.lst)

}
\keyword{datasets}
