% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors-by-waveband.r
\docType{data}
\name{electronic_components}
\alias{electronic_components}
\title{Types of sensors}
\format{
An object of class \code{character} of length 15.
}
\usage{
electronic_components
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
The sensors pointed at by this vector are those sold as electronic
components and used to construct sensors that can be deployed when interfaced
to other electronic components, possibly combined with optical filters and
entrance optics and protected inside a case or enclosure.
}
\examples{
electronic_components

}
\seealso{
\code{\link{sensors.mspct}}
}
\concept{sensors by type}
\keyword{datasets}
