% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors-mspct.r
\docType{data}
\name{image_sensors.mspct}
\alias{image_sensors.mspct}
\title{Spectral response of image sensors}
\format{
A \code{response_mspct} object containing
\code{response_spct} objects as \emph{named} members.

Spectral objects for monochrome image sensors contain two numeric variables,
with responsivity in most cases in relative energy units:
\itemize{
   \item w.length (nm)
   \item s.e.response (r.u.)
}
Spectral objects for colour image sensors contain the spectra in long form
with two numeric variables, and in addition a factor with channel ids:
\itemize{
   \item w.length (nm)
   \item s.e.response (r.u.)
   \item channel (factor with names as levels)
}
The spectral objects contain, in addition to the spectral data, metadata
stored in R attributes.
}
\usage{
image_sensors.mspct
}
\description{
A collection of response spectra for various image sensors and
digital cameras responsive to ultraviolet, visible, NIR and/or SWIR
radiation. Each spectrum in the collection contains two variables,
wavelengths (nm) at irregular intervals and spectral responsiveness (in
energy units).
Spectral data are in most cases not normalized to one at the wavelength of
maximum energy responsivity.
}
\examples{
names(image_sensors.mspct)

}
\seealso{
\code{\link[photobiology:source_spct]{response_spct}} and
\code{\link[photobiology:generic_mspct]{response_mspct}}
}
\keyword{datasets}
