% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Irradian-sensors.r
\docType{data}
\name{irradian_sensors}
\alias{irradian_sensors}
\title{Irradian Sensors}
\format{
A character vector of names of members of the collection of spectra.
}
\usage{
irradian_sensors
}
\description{
A vector of names useful for extracting subsets of spectra from
  the \code{\link{sensors.mspct}} object.
}
\details{
Irradian sensor DA211A-Cos, DA211B2-Cos, DV211Q-Cos, DV211E-Cos.

Manufacturer: Irradian
}
\note{
Spectral data digitized from figures in manufacturer's manuals.
These are typical measured data, and for specific sensor units the
calibration data supplied with the sensor must be used as individual
may differ in their spectral response. Furthermore, the digitizing
process from figures with a linear scale is prone to errors, which
are specially noticeable in regions of low responsivity.
}
\examples{
irradian_sensors

}
\references{
Manufacturer's brochure for 'DA211-v2a' not dated, downloaded on 2026-03-09.
Manufacturer's brochure for 'DV211-v2a'  not dated, downloaded on 2026-03-09.

KJ McCree. The action spectrum, absorbance and quantum yield of
  photosynthesis in crop plants. Agricultural Meteorology. 1971/72
  Vol 9, pp 191-216
}
\keyword{datasets}
