\docType{data}
\name{KIPP_PQS1_PAR_quantum}
\alias{KIPP_PQS1_PAR_quantum}
\title{Spectral data for sensors}
\description{
  Datasets containing the wavelengths (nm) at irregular
  intervals and tabulated values of spectral responsiveness
  for different sensors. Spectral response values are in
  arbitrary units.
}
\details{
  Digitized with 'enguage' from manufacturers brochures.
  This are approximate data, both because of the digitizing
  process, and because they are either typical values or
  for a particular sensor unit. Individual sensor units are
  spected to differ to some degree in spectral response.

  The variables are as follows: \itemize{ \item w.length
  (nm) \item s.e.response (a.u.)  \item s.q.respone (a.u.)
  }
}
\note{
  PQS 1 PAR Quantum Sensor

  Manufacturer: Kipp & Zonen B.V., Delftechpark 36, 2628 XH
  Delft
}
\author{
  Tähti Pohjanmies

  Pedro J. Aphalo <pedro.aphalo
}
\keyword{datasets}

