% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIPP-sensors.r
\docType{data}
\name{kipp_sensors}
\alias{kipp_sensors}
\title{Kipp Radiometers.}
\format{
An object of class \code{character} of length 6.
}
\usage{
kipp_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
Kipp CM21 pyranometer. Thermopile-detector based, class A.

Kipp CUV 5 Broadband UV Radiometer:
Detection system includes optical filters and a photodiode.

Kipp PQS 1 PAR Quantum Sensor:
Detection system includes optical filters and a photodiode.

UVS-A-T Radiometer, UVS-B-T Radiometer, Kipp UVS-E-T Erythemal Radiometer:
The detection system includes optical filters and a phosphor
that determine the spectral response. The phosphor is very
sensitive to low levels of ultraviolet radiation and is stimulated
by the UV to emit green light, which is detected by a photodiode.
The system is temperature stabilised at +25 °C to prevent
changes in spectral response and sensitivity with variations in
the ambient conditions.

Manufacturer: Kipp & Zonen B.V., Delftechpark 36, 2628 XH Delft, Nederlands.
\url{https://www.kippzonen.com/}
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{

kipp_sensors

}
\references{
Brochure 'Broadband UV Radiometers',
Brochure 'PQS 1 PAR Quantum Sensor',
Brochure 'Broadband UV Radiometers'.
Rodziewicz, T., Rajfur, M. (2019) Numerical procedures and their practical
  application in PV modules analyses. Part I: air mass. Opto-Electronics
  Review, Vol. 27, No. 1 39-57
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
