\docType{data}
\name{LI_COR_LI_210_photometric}
\alias{LI_COR_LI_210_photometric}
\title{Spectral data for sensors}
\description{
  Datasets containing the wavelengths (nm) at irregular
  intervals and tabulated values of spectral responsiveness
  for different sensors. Spectral response values are in
  arbitrary units.
}
\details{
  Digitized with 'enguage' from manufacturers brochures.
  This are approximate data, both because of the digitizing
  process, and because they are either typical values or
  for a particular sensor unit. Individual sensor units are
  spected to differ to some degree in spectral response.

  The variables are as follows: \itemize{ \item w.length
  (nm) \item s.e.response (a.u.)  \item s.q.respone (a.u.)
  }
}
\note{
  LI-210SA PHOTOMETRIIC SENSOR The LI-210SA Photometric
  Sensor utilizes a filtered silicon photodiode to provide
  a spectral response that matches the CIE curve within ±
  5% with most light sources.

  Manufacturer: LI-COR Inc., Lincoln, Nebraska
}
\author{
  Tähti Pohjanmies

  Pedro J. Aphalo <pedro.aphalo
}
\keyword{datasets}

