% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LICOR-sensors.r
\docType{data}
\name{licor_sensors}
\alias{licor_sensors}
\title{LI-COR sensors}
\format{
An object of class \code{character} of length 6.
}
\usage{
licor_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
In the LI-190SA and LI-190 PAR quantum sensor (PAR = photosynthetically
active radiation) colored glass filters are used to tailor the silicon
photodiode response to the desired quantum response. They have the same
spectral response.

The type of filter used in the LI-190R PAR quantum sensor is not described,
but improves the spectral response compared to the LI-190.

The LI-200SA features a silicon photovoltaic detector. This is not a true
'pyranometer' and should be used only in sunlight, and calibrated in
sunlight.

The LI-210SA Photometric Sensor utilizes a filtered silicon photodiode to
provide a spectral response that matches the CIE curve within ± 5% with most
light sources.

The LI-210R Photometric Sensor utilizes a filtered silicon photodiode to
provide a spectral response that matches the CIE curve.

Manufacturer: LI-COR Inc., Lincoln, Nebraska
\url{https://www.licor.com/env/}
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{
licor_sensors

}
\references{
LI-COR (2005) Instruction Manual 'LI-COR Terrestrial Radiation Sensors'.
LI-COR (2015) Technical Note 'Why Upgrade to the "R" Light and Radiation Sensors?'
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
