\docType{package}
\name{photobiologySensors-package}
\alias{photobiologySensors-package}
\title{Quantification of visible radiation for photobiology}
\description{
  Example spectral response data for some commonly used
  broad bands sensors The package is designed to complement
  package \code{photobiology}.
}
\details{
  \tabular{ll}{ Package: \tab photobiologySensors\cr Type:
  \tab Package\cr Version: \tab 0.1.0\cr Date: \tab
  2013-11-21\cr License: \tab GPL (>2.0)\cr }
}
\author{
  Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi}

  Tähti Pohjanmies (digitizing data)
}
\references{
  Aphalo, P. J., Albert, A., Björn, L. O., McLeod, A. R.,
  Robson, T. M., Rosenqvist, E. (Eds.). (2012). Beyond the
  Visible: A handbook of best practice in plant UV
  photobiology (1st ed., p. xxx + 174). Helsinki:
  University of Helsinki, Department of Biosciences,
  Division of Plant Biology. ISBN 978-952-10-8363-1 (PDF),
  978-952-10-8362-4 (paperback). Open access PDF download
  available at http://hdl.handle.net/10138/37558
}
\keyword{misc}

