% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors-by-waveband.r
\docType{data}
\name{uv_sensors}
\alias{uv_sensors}
\alias{sensors-by-waveband}
\alias{uvc_sensors}
\alias{uvb_sensors}
\alias{uva_sensors}
\alias{epar_sensors}
\alias{par_sensors}
\alias{vis_sensors}
\alias{shortwave_sensors}
\alias{red_sensors}
\alias{far_red_sensors}
\alias{blue_sensors}
\alias{erythemal_sensors}
\alias{photometric_sensors}
\alias{pyranometer_sensors}
\alias{green_sensors}
\alias{multichannel_sensors}
\title{Sensors responsive to different wavebands}
\format{
An object of class \code{character} of length 27.

An object of class \code{character} of length 2.

An object of class \code{character} of length 6.

An object of class \code{character} of length 10.

An object of class \code{character} of length 8.

An object of class \code{character} of length 9.

An object of class \code{character} of length 1.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.

An object of class \code{character} of length 4.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 11.
}
\usage{
uv_sensors

uvc_sensors

uvb_sensors

erythemal_sensors

uva_sensors

par_sensors

epar_sensors

vis_sensors

photometric_sensors

shortwave_sensors

pyranometer_sensors

red_sensors

far_red_sensors

blue_sensors

green_sensors

multichannel_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
The sensors pointed at by each of these vectors are sensitive to
  light of a certain colour within visible radiation (VIS) or to different
  bands of ultraviolet (UV) or infrared (IR) radiation.
}
\examples{
uv_sensors # ultraviolet
uvc_sensors # ultraviolet-C
uvb_sensors # ultraviolet-B
uva_sensors # ultraviolet-A
epar_sensors # extended photosynthetically active radiation
par_sensors # photosynthetically active radiation
vis_sensors # "visual" light sensors
shortwave_sensors
red_sensors
far_red_sensors
blue_sensors
multichannel_sensors

# select PAR sensors
sensors.mspct[par_sensors]

}
\seealso{
\code{\link{sensors.mspct}}
}
\concept{light sensors}
\keyword{datasets}
