% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglux-sensors.r
\docType{data}
\name{sglux_sensors}
\alias{sglux_sensors}
\title{sglux broadband sensors}
\format{
An object of class \code{character} of length 7.
}
\usage{
sglux_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
sglux SG01D-A UV-A broadband sensor (filtered SiC sensor).

sglux SG01D-B UV-B broadband sensor (filtered SiC sensor, VIS-blind).

sglux SG01D-C UV-C broadband sensor (filtered SiC sensor, "solar-blind").

sglux SG01L SiC broadband sensor (SiC sensor not filtered).

sglux TOCON blue 4 blue light broadband sensor (pre-amplified GaP detector).

TOCON preamplified sensors with similar spectral response as the diodes are
also available from sglux. The blue light sensor is only available as
preamplified TOCON. Sensors are available in different configurations with
different sensitivity and with different angular responses, and encased
to resist different environmental conditions.

Manufacturer: sglux GmbH, Berlin, Germany.
\url{https://sglux.de/en/}
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{

sglux_sensors

}
\references{
personal communication from Dr. Stefan Langer.
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
