% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skye-sensors.r
\docType{data}
\name{skye_sensors}
\alias{skye_sensors}
\title{Skye-Instruments Sensors}
\format{
An object of class \code{character} of length 11.
}
\usage{
skye_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
Skye PAR quantum Sensor SKP215, Skye PAR energy sensor SKE510, Skye Lux
sensor SKL310, Skye PAR 'Special' Sensor SKP210, Skye PAR quantum Sensor
SKP215, Skye 660/730 Sensor (Red/Far Red) Sensor SKR110 (red and far-red
channels), Skye Pyranometer Sensor SKS1110, Skye UV-A Sensor Sensor SKU421,
Skye UV-A Sensor Sensor SKU421 V. 3, Skye UV-B Sensor Sensor SKU430 V. 3,
Skye UVI Sensor Sensor SKU440 V. 3.

Manufacturer: Skye Instruments, Llandrindod Wells, Powys, UK.
\emph{As of 2026, out of business.}
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{
skye_sensors

}
\references{
Manufacturer's brochure for 'SKP215' dated 2007-10-03.
Manufacturer's brochure for 'SKP510' dated 2007-10-02.
Manufacturer's brochure for 'SKL310' dated 2007-10-09
Manufacturer's brochure for 'SKP210' dated 2007-10-03
Manufacturer's brochure for 'Skye 660/730 Sensor (Red/Far Red) Sensor SKR110' dated 2007-10-09
Manufacturer's brochure 'Skye Pyranometer Sensor SKS1110' dated 2009-08-19
Manufacturer's 'UV-A, UV-B & UV-I Sensors' Manual, Iss. 1.1
Manufacturer's 'SKU 421 UVA Sensor' V. 3 brochure, not dated, dowloaded on 2015-01-24
Manufacturer's 'SKU 430 UVB Sensor' V. 3 brochure, not dated, dowloaded on 2015-01-24
Manufacturer's 'SKU 440 UV Index Sensor' V. 3 brochure, not dated, dowloaded on 2015-01-24

KJ McCree. The action spectrum, absorbance and quantum yield of
  photosynthesis in crop plants. Agricultural Meteorology. 1971/72
  Vol 9, pp 191-216
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
