\docType{data}
\name{TOCON_E2_erythemal}
\alias{TOCON_E2_erythemal}
\title{Spectral data for sensors}
\description{
  Datasets containing the wavelengths (nm) at irregular
  intervals and tabulated values of spectral responsiveness
  for different sensors. Spectral response values are in
  arbitrary units.
}
\details{
  Digitized with 'enguage' from manufacturers brochures.
  This are approximate data, both because of the digitizing
  process, and because they are either typical values or
  for a particular sensor unit. Individual sensor units are
  spected to differ to some degree in spectral response.

  The variables are as follows: \itemize{ \item w.length
  (nm) \item s.e.response (a.u.)  \item s.q.respone (a.u.)
  }
}
\note{
  TOCON_E2 (TOCON_ERYCA) Pre-amplified SiC UV-Index
  Photodetector Manufacturer: sglux GmbH, Max-Planck-Str.
  3, D-12489 Berlin
}
\author{
  Tähti Pohjanmies

  Pedro J. Aphalo <pedro.aphalo
}
\keyword{datasets}

