% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavebands.r
\docType{data}
\name{uv_sensors}
\alias{uv_sensors}
\alias{uvc_sensors}
\alias{uvb_sensors}
\alias{erythemal_sensors}
\alias{uva_sensors}
\alias{par_sensors}
\alias{epar_sensors}
\alias{vis_sensors}
\alias{photometric_sensors}
\alias{shortwave_sensors}
\alias{pyranometer_sensors}
\alias{red_sensors}
\alias{far_red_sensors}
\alias{blue_sensors}
\alias{green_sensors}
\alias{multichannel_sensors}
\alias{electronic_components}
\title{Sensors responsive to different wavebands}
\format{
A vector of character strings.

An object of class \code{character} of length 2.

An object of class \code{character} of length 6.

An object of class \code{character} of length 8.

An object of class \code{character} of length 7.

An object of class \code{character} of length 8.

An object of class \code{character} of length 1.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 10.

An object of class \code{character} of length 9.
}
\usage{
uv_sensors

uvc_sensors

uvb_sensors

erythemal_sensors

uva_sensors

par_sensors

epar_sensors

vis_sensors

photometric_sensors

shortwave_sensors

pyranometer_sensors

red_sensors

far_red_sensors

blue_sensors

green_sensors

multichannel_sensors

electronic_components
}
\description{
Vectors of names of members of the collection of spectra
\code{\link{sensors.mspct}} by the wavelengths or colours they are sensitive
to.
}
\examples{
uv_sensors # ultraviolet
uvc_sensors # ultraviolet-C
uvb_sensors # ultraviolet-B
uva_sensors # ultraviolet-A
epar_sensors # extended photosynthetically active radiation
par_sensors # photosynthetically active radiation
vis_sensors # "visual" light sensors
shortwave_sensors
red_sensors
far_red_sensors
blue_sensors
multichannel_sensors
electronic_components # native spectral response, sold as parts

# select PAR sensors
sensors.mspct[par_sensors]

}
\seealso{
\code{\link{sensors.mspct}}
}
\concept{sensors by waveband or color}
\keyword{datasets}
