% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalTech-sensors.r
\docType{data}
\name{vitaltech_sensors}
\alias{vitaltech_sensors}
\title{Vital Technologies sensors}
\format{
A character vector of names of members of the collection of spectra.
}
\source{
Data kindly made available by Lasse Ylianttila.
}
\usage{
vitaltech_sensors
}
\description{
A vector of names useful for extracting subsets of spectra from
  the \code{\link{sensors.mspct}} object.
}
\details{
Vital "Blue Wave" BW-20 UV-B "erythemal" radiometer. Data for a specific
unit included in a sensor intercomparison event.

Manufacturer: Vital Technologies, Canada. Company no longer in business.
}
\examples{

vitaltech_sensors

}
\references{
Leszczynski K, Jokela K, Ylianttila L, Visuri R, Blumthaler M. 1997. Report of the WMO/STUK Intercomparison
of erythemally-weighted solar UV radiometers (Spring/Summer 1995, Helsinki, Finland). WMO-GAW Report No. 112, 90 pages.

Leszczynski K, Jokela K, Ylianttila L, Visuri R, Blumthaler M. 1998. Erythemally weighted radiometers in solar
UV monitoring: results from the WMO/STUK Intercomparison. Photochem. Photobiol. 67(2):212-221.
}
\keyword{datasets}
