% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytik-sensors.r
\docType{data}
\name{analytik_sensors}
\alias{analytik_sensors}
\title{Analytik-Jena sensors.}
\format{
A character vector of names of members of the collection of spectra.
}
\usage{
analytik_sensors
}
\description{
A vector of names useful for extracting subsets of spectra from
  the \code{\link{sensors.mspct}} object.
}
\details{
UVX-25, UVX-31 and UVX-36 sensor for UV-C, UV-B and UV-A radiation,
respectively. Detection system includes optical filters and a Si photodiode.
}
\note{
Digitized with 'DigitizeIt' from figures 6, 7 and 8 from manufacturers
product manual. This are approximate data, both because of the digitizing
process, and because they are either typical values or for a particular
sensor unit. Individual sensor units are expected to differ to some degree in
spectral response.

Manufacturer: Analytik Jena US LLC, Upland, CA, USA (former UVP)
\url{https://www.analytik-jena.com/} or \url{https://www.uvp.com/}.
}
\examples{

analytik_sensors

}
\references{
Product manual: 'UVP UVX Radiometer: Instruction Guide'
}
\keyword{datasets}
