% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berger-sensors.r
\docType{data}
\name{berger_sensors}
\alias{berger_sensors}
\title{'Berger' UV-Biometer}
\format{A character vector of members of the collection of spctra.}
\usage{
berger_sensors
}
\description{
A vector of indeces for stracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\note{
'Berger' UV-Biometer (Fig. 1 in Berger, 1994).
Digitized with 'enguage' from manufacturers brochures. These are approximate
data, both because of the digitizing process, and because they are
either typical values or for a particular sensor unit. Individual sensor units
are expected to differ to some degree in spectral response.
}
\examples{
berger_sensors

}
\references{
Berger, Daniel (n.d., ca. 1994) A comparison of Spectroradiometers to Radiometers for
UV Radiation Measurements. Solar Light Co., Inc., Philadelphia.
}
\keyword{datasets}
