% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaT-sensors.r
\docType{data}
\name{deltat_sensors}
\alias{deltat_sensors}
\title{Delta-T sensors}
\format{
A character vector of names of members of the collection of spectra.
}
\usage{
deltat_sensors
}
\description{
A vector of names useful for extracting subsets of spectra from
  the \code{\link{sensors.mspct}} object.
}
\note{
Spectral data digitized from figure in manufacturer's manual.
These are typical measured data, and for specific sensor units the
calibration data supplied with the sensor must be used as individual
units may differ in their spectral response. Furthermore, the digigitizing
process from figures with a linear scale is prone to errors, which
are specially noticeable in regions of low responsivity.

Manufacturer: Delta-T Devices Ltd, CAMBRIDGE CB25 0EJ, UK
\url{https://delta-t.co.uk/}
}
\examples{
deltat_sensors

}
\references{
Manufacturer's User Manual Version: 1.0 dated Nov 2010
}
\keyword{datasets}
