% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors-by-waveband.r
\docType{data}
\name{electronic_components}
\alias{electronic_components}
\title{Types of sensors}
\format{
An object of class \code{character} of length 15.
}
\usage{
electronic_components
}
\description{
Types of sensors
}
\examples{
electronic_components

}
\concept{sensors by type}
\keyword{datasets}
