% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-sensors.r
\docType{data}
\name{ideal_sensors}
\alias{ideal_sensors}
\title{Idealized sensors}
\format{
An object of class \code{character} of length 2.
}
\usage{
ideal_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
Flat spectral response to spectral energy irradiance and to spectral photon
irradiance.
}
\note{
Exact values based on physical principles.
}
\examples{
ideal_sensors

}
\references{
Data sheet for TSL254R (TAOS, TAOS071C - SEPTEMBER 2007)

Data sheet for TSL257 (v1-00, 2016-Jul-25)

Data sheet for TSL2591 (ams DS000338, v3-00, 2023-Feb-08)

Data sheet for AS7263 (ams DS000476, v4-00, 2022-Nov-30)

Data sheet for AS7331 (ams DS001047, v4-00, 2023-Mar-24)

Data sheet for AS7341 (ams DS000504, v3-00, 2020-Jun-25)

Data sheet for AS7343 (ams DS001046, v6-00, 2023-Jun-07)

Data sheet for BPX65 (Version 1.4, 2023-01-24)
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
