% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-sensors.r
\docType{data}
\name{ideal_sensors}
\alias{ideal_sensors}
\title{Idealized sensors}
\format{
A character vector of names of members of the collection of spectra.
}
\usage{
ideal_sensors
}
\description{
A vector of names useful for extracting subsets of spectra from
  the \code{\link{sensors.mspct}} object.
}
\note{
Flat spectral response to spectral energy irradiance and to spectral photon
irradiance.
}
\examples{
ideal_sensors

}
\keyword{datasets}
