% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors-mspct.r
\docType{data}
\name{image_sensors.mspct}
\alias{image_sensors.mspct}
\title{Spectral response of image sensors}
\format{
A \code{response_mspct} object containing
\code{response_spct} objects as \emph{named} members.

Spectral objects for monochrome image sensors contain two numeric variables,
with quantum efficiency in some cases in relative energy units:
\itemize{
   \item w.length (nm)
   \item s.q.response (r.u.)
}
Spectral objects for colour image sensors contain the spectra in long form
with two numeric variables, and in addition a factor with channel ids:
\itemize{
   \item w.length (nm)
   \item s.q.response (r.u.)
   \item channel (factor with names as levels)
}
The spectral objects contain, in addition to the spectral data, metadata
stored in R attributes \code{sensor.properties}, \code{what.measured},
\code{how.measured}, and \code{comment}.
}
\usage{
image_sensors.mspct
}
\description{
A collection of response spectra for various image sensors and
digital cameras responsive to ultraviolet, visible, NIR and/or SWIR
radiation. Each spectrum in the collection contains two variables,
wavelengths (nm) at irregular intervals and spectral responsiveness (in
photon units = quantum efficiency).
Spectral data are normalized to one at the wavelength of
maximum quantum efficiency only when this is how the original source
published the data.
}
\examples{
names(image_sensors.mspct)

}
\seealso{
\code{\link[photobiology:source_spct]{response_spct}} and
\code{\link[photobiology:generic_mspct]{response_mspct}}
}
\keyword{datasets}
