% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Irradian-sensors.r
\docType{data}
\name{irradian_sensors}
\alias{irradian_sensors}
\title{Irradian Sensors}
\format{
An object of class \code{character} of length 4.
}
\usage{
irradian_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
Irradian sensors DA211A-Cos (UV-A), DA211B2-Cos (UV-B), DV211Q-Cos (PAR),
DV211E-Cos (pyranometer).

Manufacturer: Irradian Ltd., Tranent, East Lothian, Scotalnd, U.K.
}
\note{
Digitized with 'DigitizeIt' or 'engauge' from manufacturers publications.
This are approximate data, both because of the digitizing process, and
because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some degree in spectral
response. In the case of multichannel sensors, the spectra for all channels
are plotted in the same figure in the data sheets and in out-of-band regions
the lines overlap making accurate extraction of values difficult. Variation
from unit to unit affects both response and center/peak wavelengths.
}
\examples{
irradian_sensors

}
\references{
Manufacturer's brochure for 'DA211-v2a' not dated, downloaded on 2026-03-09.
Manufacturer's brochure for 'DV211-v2a'  not dated, downloaded on 2026-03-09.

KJ McCree. The action spectrum, absorbance and quantum yield of
  photosynthesis in crop plants. Agricultural Meteorology. 1971/72
  Vol 9, pp 191-216
}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
