% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiesclima-sensors.r
\docType{data}
\name{thiesclima_sensors}
\alias{thiesclima_sensors}
\title{Thies Clima sensors}
\format{A character vector of members of the collection of spctra.}
\source{
Digitized with 'DigitizeIt' from manufacturers manual. These are approximate
data, both because of the digitizing process, and because they are
either typical values or for a particular sensor unit. Individual sensor units
are expected to differ to some degree in spectral response.
}
\usage{
thiesclima_sensors
}
\description{
A vector of indeces for stracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
Thies Clima E1.c broadband UVB sensor

Manufacturer: Thies Clima, Göttingen, DE
 \url{http://www.thiesclima.com/}
}
\examples{

thiesclima_sensors

}
\keyword{datasets}
