% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavebands.r
\docType{data}
\name{uv_sensors}
\alias{uv_sensors}
\alias{'sensors}
\alias{by}
\alias{waveband'}
\alias{uvc_sensors}
\alias{uvb_sensors}
\alias{erythemal_sensors}
\alias{uva_sensors}
\alias{par_sensors}
\alias{vis_sensors}
\alias{photometric_sensors}
\alias{shortwave_sensors}
\alias{pyranometer_sensors}
\alias{red_sensors}
\alias{far_red_sensors}
\alias{blue_sensors}
\alias{multichannel_sensors}
\title{Sensors responsive to different wavebands}
\format{A vector of character strings.}
\usage{
uv_sensors

uvc_sensors

uvb_sensors

erythemal_sensors

uva_sensors

par_sensors

vis_sensors

photometric_sensors

shortwave_sensors

pyranometer_sensors

red_sensors

far_red_sensors

blue_sensors

multichannel_sensors
}
\description{
Names of datasets containing the wavelengths and tabulated values spectral emittance
for broadband sensors from various suppliers.
}
\examples{
uv_sensors
uvc_sensors
uvb_sensors
uva_sensors
par_sensors
vis_sensors
shortwave_sensors
red_sensors
far_red_sensors
blue_sensors
multichannel_sensors

# select PAR sensors
sensors.mspct[par_sensors]

}
\seealso{
\code{\link{sensors.mspct}}
}
\keyword{datasets}
