% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vishay-sensors.r
\docType{data}
\name{vishay_sensors}
\alias{vishay_sensors}
\title{'Vishay' sensors}
\format{
A character vector of names of members of the collection of spectra.
}
\usage{
vishay_sensors
}
\description{
A vector of indexes for extracting subsets of spectra from the
\code{\link{sensors.mspct}} object.
}
\details{
Vishay VEML6075 UVA and UVB sensor with I2C digital interface. Two spectra,
one for each channel.
}
\note{
Digitized with 'DigitizeIt' from manufacturers' electronic component data
sheet. This are approximate data, both because of the digitizing process,
and because they are either typical values or for a specific sensor unit.
Individual sensor units are expected to differ to some
degree in spectral response.

Manufacturer: VISHAY INTERTECHNOLOGY, INC. Shelton, CT, USA. \url{https://www.vishay.com/}.
}
\examples{

vishay_sensors

}
\references{
Document Number: 84304, Rev. 1.2, 23-Nov-16; VISHAY INTERTECHNOLOGY, INC.
}
\keyword{datasets}
